Initial Commit
A more or less fully functional app.

This time around, we're only doing time based slices. I got space on an
MCS server, so I can massively speed up the time it takes to transfer
the tarballs.

There's also a new suite of slice'n'dice tools from the Waggle team that
makes the job even faster.

The MCS box will do all the data slicing and then upload it S3. The
daily job will also upload a tarball of metadata to S3 for the app to
consume and update itself.

In all honesty, this probably doesn't even need to be a real app; I'm
just too lazy to design this as flat pages that are updated every day.